refactor(view): extract layout (renderMainLayout, contentInnerWidth, renderTreePanel) into view_layout.go; leave full-screen helpers in place; no behavior change
